+++ packages/forma-36-react-components/src/components/Card/CardActions/CardActions.tsx
@@ -71,6 +71,7 @@ export class CardActions extends Component<
         position="bottom-right"
         className={className}
         isOpen={this.state.isDropdownOpen}
+        testId={testId}
         toggleElement={
           <IconButton
             iconProps={{ icon: 'MoreHorizontal' }}