+++ packages/components/forma-36-react-timepicker/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-alpha.30](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-timepicker@0.2.0-alpha.29...@contentful/forma-36-react-timepicker@0.2.0-alpha.30) (2020-01-20)
+
+**Note:** Version bump only for package @contentful/forma-36-react-timepicker
+
+
+
+
+
+# [0.2.0-alpha.29](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-timepicker@0.2.0-alpha.28...@contentful/forma-36-react-timepicker@0.2.0-alpha.29) (2020-01-16)
+
+**Note:** Version bump only for package @contentful/forma-36-react-timepicker
+
+
+
+
+
+# [0.2.0-alpha.28](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-timepicker@0.2.0-alpha.27...@contentful/forma-36-react-timepicker@0.2.0-alpha.28) (2020-01-16)
+
+**Note:** Version bump only for package @contentful/forma-36-react-timepicker
+
+
+
+
+
+# [0.2.0-alpha.27](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-timepicker@0.2.0-alpha.26...@contentful/forma-36-react-timepicker@0.2.0-alpha.27) (2020-01-14)
+
+**Note:** Version bump only for package @contentful/forma-36-react-timepicker
+
+
+
+
+
 # [0.2.0-alpha.26](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-timepicker@0.2.0-alpha.25...@contentful/forma-36-react-timepicker@0.2.0-alpha.26) (2020-01-09)
 
 **Note:** Version bump only for package @contentful/forma-36-react-timepicker